Improved project page views and URLs
The handling now checks the "api-id" property for a project's API ID.
This allows more human-friendly URLs for projects that have files in the
CMS (like N-ICE2015). 301 redirects have been implemented so that if a
project file exists in the CMS, it is this file's URL that is exposed.
The master template (npweb.jsp) has had all logic pertaining to handling
of projects removed.