Add regular common.js (not JSP) w/ new Highslide handling
The new commons.js is a copy of the current commons.js.jsp – except for
Highslide config, which is changed only in terms of how it is handled.
This new approach enables HS only when necessary, via a call made from
the master template, in which HS asset URIs and a language string is
passed.