Update "search panel", new "toggleables", centralized filtering/pagination/testing
This is a large bulk update. The new `.search-panel` replaces the old
`.searchbox-big`. The search panel will typically contain a search box,
and often times filtering as well. For this, common components are
introduced, e.g. the `.searchbox` and the more complex component wrapped
in `.search-panel__filters`. A new general approach to toggleable
content is also included: a `.toggler` controller, directly followed by
a `.toggleable` container for the expandable/collapsible content.
Testing of API endpoints (services) is updated to use the centralized
version in `no.npolar.common.data.api.util.APIUtil`.